Refactored callbackqueryhandler, fixed minor bugs,removed useless functions (#65)
