+++ .gitignore
@@ -124,11 +124,6 @@ config/drive_credentials.json
 config/settings.yaml
 data/DMI_DB.db
 module/*.pyc
-data/archivioAvvisi.dat
-data/avvisiInSospeso.dat
-data/avviso.dat
 pydrive
 pydrive/*
-data/*.xls
-debugDrive.txt
 .vscode
